Move fetchBackendCatalog into the common tools.ts
